backend/requirements.txt
fastapi[standard]
uvicorn[standard]
google-cloud-storage
PyYAML
xgboost
pandas
shap
numpy
pandas